adding a function in each file
